requirements.txt
requests>=2.31.0
websocket-client>=1.7.0
python-dotenv>=1.0.0
rich>=13.7.0
textual>=0.47.0
prompt_toolkit>=3.0.43
tabulate>=0.9.0
colorama>=0.4.6
aiohttp>=3.9.0
asyncio-nats-client>=0.11.4
flask>=3.0.0
flask-cors>=4.0.0
apscheduler>=3.10.4
numpy>=1.26.0
